FinalProject/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
Django==3.1.4
django-crispy-forms==1.10.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
Faker==5.1.0
idna==2.10
lxml==4.6.2
mysql-connector==2.2.9
mysql-connector-python==8.0.22
Pillow==8.0.1
pkg-resources==0.0.0
protobuf==3.14.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
reverse==0.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
text-unidecode==1.3
todo==0.1
urllib3==1.26.2
whitenoise==5.2.0